feat: (1) map processed proposals
